# open-api/swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    { "url": "http://localhost:4004", "description": "", "variables": {} }
  ],
  "info": {
    "version": "v1.0",
    "title": "be-challenge",
    "description": "",
    "termsOfService": "",
    "contact": {},
    "license": { "name": "MIT" }
  },
  "paths": {
    "/api/v1/video/upload": { 
      "post": {
        "summary": "upload video",
        "operationId": "uploadvideo",
        "parameters": [],
        "responses": { "200": { "description": "", "headers": {} } },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              },
              "example": [
                {
                  "key": "samplefile",
                  "type": "file",
                  "src": "/Users/manavshrivastava/Desktop/k8s.png"
                }
              ]
            }
          }
        }
      }
    },
...
    "/api/v1/course": {
      "post": {
        "summary": "create course",
        "operationId": "createcourse",
        "parameters": [],
        "responses": { "200": { "description": "", "headers": {} } },
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": {
                  "title": "course 1",
                  "subjects": [
                    "60a162312d94a16058f0fdd4",
                    "60a162312d94a16058f0fdd4"
                  ]
                }
              },
              "example": "{\n    \"title\":\"course 1\",\n    \"subjects\":[\"60a162312d94a16058f0fdd4\",\n    \"60a162312d94a16058f0fdd4\"]\n}"
            }
          }
        }
      }
    },
...
    "/api/v1/subject": {
      "post": {
        "summary": "create subject",
        "operationId": "createsubject",
        "parameters": [],
        "responses": { "200": { "description": "", "headers": {} } },
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": {
                  "title": "subbject 1",
                  "VideoWebinars": [
                    "60a162312d94a16058f0fdd4",
                    "60a162312d94a16058f0fdd4"
                  ]
                }
              },
              "example": "{\n    \"title\":\"subbject 1\",\n    \"VideoWebinars\":[\"60a162312d94a16058f0fdd4\",\n    \"60a162312d94a16058f0fdd4\"]\n}"
            }
          }
        }
      }
    }
  },
...
  "security": [],
...
}